requirements.txt
addict==2.4.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
babel==2.17.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
csvw==3.5.1
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.11
dlinfo==2.0.0
docopt==0.6.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
espeakng-loader==0.2.4
fastapi==0.115.14
ffmpy==0.6.0
filelock==3.18.0
fsspec==2025.5.1
gradio==5.35.0
gradio_client==1.10.4
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.1
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kokoro==0.9.4
langcodes==3.5.0
language-tags==1.2.0
language_data==1.3.0
llvmlite==0.44.0
loguru==0.7.3
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
misaki==0.9.4
more-itertools==10.7.0
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.5
num2words==0.5.14
numba==0.61.2
numpy==2.2.6
openai-whisper==20250625
orjson==3.10.18
packaging==25.0
pandas==2.3.0
phonemizer-fork==3.3.2
pillow==11.2.1
preshed==3.0.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rfc3986==1.5.0
rich==14.0.0
rpds-py==0.25.1
ruff==0.12.1
sacremoses==0.1.1
safehttpx==0.1.6
safetensors==0.5.3
segments==2.3.0
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soundfile==0.13.1
spacy==3.8.7
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.2
sympy==1.14.0
thinc==8.3.6
tiktoken==0.9.0
tokenizers==0.21.2
tomlkit==0.13.3
torch==2.7.1
tqdm==4.67.1
transformers==4.53.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.34.3
wasabi==1.1.3
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.2